feat: refactor roles management in admin UI with new RolesView component and related files
